Moving credit cards endpoints to frontend
