+++ src/app/helpers/sortBy.ts
@@ -1,0 +1,3 @@
+export const sortBy = (key: string) => {
+  return (a: any, b: any) => (a[key] > b[key] ? 1 : b[key] > a[key] ? -1 : 0);
+};
