+++ src/app/interfaces/creditCard/creditCard.ts
@@ -1,3 +1,36 @@
+import { CIRCLE_FAILURE_CODES } from 'app/constants/contants';
+
+export type Status = 'approved' | 'pending' | 'rejected' | 'needsDepositVerification';
+export type DepositVerificationStatus =
+  | 'notStarted'
+  | 'pending'
+  | 'submissionFailed'
+  | 'awaitingVerification'
+  | 'successful'
+  | 'failed';
+
+export interface ICreditCardsResponse {
+  result: CreditCardData[];
+  success: boolean;
+}
+
+export interface ICreditCardResponse {
+  result: CreditCardData;
+  success: boolean;
+}
+
+export interface CreditCardData {
+  id: string;
+  name: string;
+  time: string;
+  billingInfo: Record<string, string>;
+  status: Status;
+  errorCode: string;
+  depositVerificationStatus: DepositVerificationStatus;
+  depositVerificationErrorCode: keyof typeof CIRCLE_FAILURE_CODES;
+  data: Record<string, string> | null;
+}
+
 export interface ICreditCardError {
   data: CreditCardError;
   status: number;
